adding unit tests

+++ model/forum/data.php
@@ -12,11 +12,13 @@ class Data extends Forum {
     public function create() {
         if ( empty( in('session_id') ) ) return error( ERROR_SESSION_ID_EMPTY );
         if( empty( in('config_idx') ) ) return error( ERROR_FORUM_CONFIG_IDX_EMPTY );
-
+        if( ! is_numeric( in('config_idx') ) ) return error( ERROR_CONFIG_IDX_NOT_NUMBER );
         $config = forum_config()->getConfig();
         if( empty( $config ) ) return error( ERROR_FORUM_CONFIG_NOT_EXIST );
         $user = user()->load_by_session_id( in('session_id') );
         if ( empty($user) ) return error( ERROR_USER_NOT_EXIST );
+        if( empty( in('user_idx') ) ) return error( ERROR_USER_IDX_EMPTY );
+        if( !is_numeric( in('user_idx') ) ) return error( ERROR_USER_IDX_NOT_NUMBER);
 
         $data = [];
         $data['user_idx'] = in('user_idx');
@@ -24,12 +26,18 @@ class Data extends Forum {
         $data['content'] = in('content');
         $data['config_idx'] = in('config_idx');
         if( empty( $data['title'] ) ) return error( ERROR_FORUM_DATA_TITLE_EMPTY );
+        if( strlen( $data['title'] ) > 256 ) return error( ERROR_TITLE_TOO_LONG );
         $forumdata_idx = $this->insert( $data );
         if ( $forumdata_idx <= 0 ) return error( $forumdata_idx );
         success( ['forum_data'=>$forumdata_idx] );
     }
     public function edit() {
         if( empty( in('session_id') ) ) return error( ERROR_SESSION_ID_EMPTY );
+//        if( empty( in('user_idx') ) ) return error( ERROR_IDX_EMPTY );
+        if( !is_numeric(in('user_idx') ) ) return error( ERROR_USER_IDX_NOT_NUMBER );
+        if( strlen( in('title') ) > 256 ) return error( ERROR_TITLE_TOO_LONG );
+        $user = user()->load_by_session_id( in('session_id') );
+        if( empty( $user ) ) return error( ERROR_USER_NOT_EXIST );
         $data = [];
         $data['idx'] = in('idx');
         $data['title'] = in('title');
@@ -43,9 +51,12 @@ class Data extends Forum {
     }
     public function delete( $compatibility_arg1 = null ) {
         $idx = in('idx');
+        if( empty( in('session_id') ) ) return error( ERROR_SESSION_ID_EMPTY );
         if ( empty($idx) ) $idx = "id='".in('id')."'"; // if it is not string, it is id.
         $data = $this->load( $idx );
+        $user = user()->load_by_session_id( in('session_id') );
         if( !$data ) error( ERROR_FORUM_DATA_NOT_EXIST );
+        if( $data['user_idx'] != $user['idx'] ) return error( ERROR_USER_IDX_NOT_MATCHED );
         $this->destroy();
         success();
     }
